template/tests: factor shared _load_impl_module into tests/common.py; update tests to use it with robust sys.path handling
