+++ reproduction_template/tests/test_smoke.py
@@ -1,17 +1,14 @@
+import pathlib
+import sys
+
+# Ensure this tests directory is on sys.path to import shared helper
+_TESTS_DIR = pathlib.Path(__file__).parent
+if str(_TESTS_DIR) not in sys.path:
+    sys.path.insert(0, str(_TESTS_DIR))
+
+from common import _load_impl_module
+
+
 def test_placeholder():
     # Basic smoke test to ensure template imports and runs
-    import importlib.util
-    import pathlib
-    import sys
-
-    impl = pathlib.Path(__file__).resolve().parents[1] / "implementation.py"
-    assert impl.exists(), "implementation.py missing"
-
-    # Ensure local 'lib' package is importable
-    sys.path.insert(0, str(impl.parent))
-    spec = importlib.util.spec_from_file_location("impl", impl)
-    assert spec is not None
-    module = importlib.util.module_from_spec(spec)
-    assert module is not None
-    assert spec.loader is not None
-    spec.loader.exec_module(module)
+    _ = _load_impl_module()
